+++ R/LLMR_tidy.R
@@ -454,7 +454,7 @@ llm_fn <- function(x,
 #'
 #' # Multimodal: include an image path with role 'file'
 #' pics <- tibble::tibble(
-#'   img    = c("inst/extdata/cat.png", "inst/extdata/dog.jpg"),
+#'   img    = c("path/to/cat.png", "path/to/dog.jpg"),
 #'   prompt = c("Describe the image.", "Describe the image.")
 #' )
 #' pics |>